// web/src/api/generated/model/leaveDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { LeaveKind } from './leaveKind';

export interface LeaveDto {
  id: string;
  employeeId: string;
  kind: LeaveKind;
  from: string;
  to: string;
  days: number;
  /** @nullable */
  notes: string | null;
}