// web/src/api/generated/model/leaveInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { LeaveKind } from './leaveKind';

export interface LeaveInput {
  employeeId: string;
  kind: LeaveKind;
  from: string;
  to: string;
  /** @nullable */
  days: number | null;
  /** @nullable */
  notes: string | null;
}