// web/src/api/generated/model/salaryComponentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Baba.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ComponentCalculation } from './componentCalculation';
import type { SalaryComponentKind } from './salaryComponentKind';

export interface SalaryComponentInput {
  code: string;
  nameAr: string;
  nameEn: string;
  kind: SalaryComponentKind;
  calculation: ComponentCalculation;
  defaultValue: number;
  /** @nullable */
  accountId: string | null;
  isInsurable: boolean;
  inEndOfService: boolean;
}